chore(functions): rename OpenAI naming to LLM (no behavior change)
